requirements.txt
absl-py==0.12.0
accelerate==0.26.1
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
alabaster==0.7.12
ale-py==0.7.4
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.7.2
astroid==2.4.2
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.3
asyncer==0.0.2
attrs==21.4.0
AutoROM==0.4.2
AutoROM.accept-rom-license==0.4.2
Babel==2.9.0
backcall==0.2.0
backoff==2.2.1
backports.zoneinfo==0.2.1
bcrypt==4.0.1
bidict==0.22.1
bitsandbytes==0.42.0
black==22.3.0
bleach==3.3.0
blessed==1.20.0
blis==0.7.11
Brotli==1.0.9
build==1.0.3
cachetools==4.2.1
casadi==3.6.3
catalogue==2.0.10
certifi==2020.12.5
cffi==1.14.5
chainlit==1.0.101
chardet==4.0.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.3
cloudpathlib==0.16.0
cloudpickle==2.0.0
cmasher==1.6.3
codefind==0.1.3
coleo==0.3.2
colorama==0.4.4
coloredlogs==15.0.1
colorspacious==1.1.2
confection==0.1.4
contourpy==1.0.7
coverage==6.3.2
cryptography==40.0.1
ctransformers==0.2.27
cycler==0.10.0
cymem==2.0.8
Cython==0.29.26
dash==2.6.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.5.14
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.8
Django==4.2.13
docutils==0.17.1
e13tools==0.9.6
entrypoints==0.3
etils==0.8.0
exceptiongroup==1.2.0
execnet==1.9.0
executing==1.2.0
fabric==2.7.1
faiss-cpu==1.7.4
fastai==2.7.15
fastapi==0.100.1
fastapi-socketio==0.0.10
fastcore==1.5.33
fastdownload==0.0.7
fasteners==0.17.3
fastprogress==1.0.3
filelock==3.13.1
filetype==1.2.0
flake8==4.0.1
flake8-bugbear==22.4.25
Flask==2.2.2
Flask-Compress==1.13
flatbuffers==1.12
fluidnet-cpp==0.0.0
fonttools==4.39.4
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.2
gast==0.3.3
giving==0.4.2
glfw==2.5.3
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
googleapis-common-protos==1.62.0
graphlib-backport==1.0.3
greenlet==3.0.3
grpcio==1.60.0
gym==0.21.0
gym-notices==0.0.6
h11==0.14.0
h5py==2.10.0
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.1
huggingface-hub==0.20.2
humanfriendly==10.0
idna==2.10
imageio==2.18.0
imageio-ffmpeg==0.4.8
imagesize==1.3.0
importlab==0.7
importlib-metadata==6.11.0
importlib-resources==5.7.1
iniconfig==1.1.1
inquirerpy==0.3.4
install==1.3.5
invoke==1.7.3
ipykernel==5.5.0
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.5.3
itsdangerous==2.1.2
jacobi==0.8.1
jax==0.3.21
jaxlib==0.3.20
jedi==0.18.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.5
jsonpatch==1.33
jsonpointer==2.4
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.5.1
jupyterlab==3.0.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kubernetes==29.0.0
langchain==0.1.0
langchain-community==0.0.11
langchain-core==0.1.10
langcodes==3.4.0
langsmith==0.0.80
language_data==1.2.0
Lazify==0.4.0
lazy-object-proxy==1.4.3
libcst==0.4.1
literalai==0.0.102
livereload==2.6.3
llama2==0.0.1.dev0
marisa-trie==1.1.0
Markdown==3.3.4
MarkupSafe==2.1.1
marshmallow==3.20.2
matplotlib==3.7.1
mccabe==0.6.1
milatools==0.0.18
mistune==0.8.4
mmh3==4.1.0
monotonic==1.6
moviepy==1.0.3
mpmath==1.3.0
mujoco-py==2.1.2.14
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==0.4.3
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.8
networkx==2.8
ninja==1.10.2.3
nltk==3.8.1
notebook==6.3.0
numpngw==0.1.0
numpy==1.24.4
oauthlib==3.2.2
onnxruntime==1.16.3
opencv-python==4.10.0.84
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.2
Panda3D==1.10.10
panda3d-gltf==0.13
panda3d-simplepbr==0.9
pandas==1.4.2
pandocfilters==1.4.3
paramiko==3.1.0
parso==0.8.1
pathlib2==2.3.7.post1
pathspec==0.9.0
pexpect==4.8.0
pfzy==0.3.4
phiflow @ git+https://github.com/tum-pbs/PhiFlow@18db0928524754d4411bd72eca74dceac499305e
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==4.2.1
plotly==5.10.0
pluggy==1.0.0
posthog==3.3.1
preshed==3.0.9
proglog==0.1.10
prometheus-client==0.9.0
prompt-toolkit==3.0.18
protobuf==4.25.2
psutil==5.9.0
ptera==1.4.1
ptyprocess==0.7.0
pulsar-client==3.4.0
py==1.11.0
py-cpuinfo==9.0.0
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.2.5
pycodestyle==2.8.0
pycparser==2.20
pydantic==2.5.3
pydantic_core==2.14.6
pyenchant==3.2.2
pyflakes==2.4.0
Pygments==2.8.1
PyJWT==2.8.0
pylint==2.6.0
PyNaCl==1.5.0
pyobjc-core==8.1
pyobjc-framework-Cocoa==8.1
pyparsing==2.4.7
pypdf==3.17.4
pyperclip==1.8.2
PyPika==0.48.9
pyproject_hooks==1.0.0
pyrsistent==0.17.3
pytest==7.1.2
pytest-cov==3.0.0
pytest-env==0.6.2
pytest-forked==1.4.0
pytest-xdist==2.5.0
python-dateutil==2.8.1
python-dotenv==1.0.0
python-engineio==4.8.2
python-graphql-client==0.4.3
python-multipart==0.0.6
python-socketio==5.11.0
pytype==2022.4.26
pytz==2021.1
PyYAML==6.0
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
questionary==1.10.0
reactivex==4.0.4
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.0
rsa==4.7.2
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.8.0
screeninfo==0.8
seaborn==0.13.2
Send2Trash==1.5.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
simple-websocket==1.0.0
six==1.15.0
smart-open==6.4.0
sniffio==1.2.0
snowballstemmer==2.2.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==4.5.0
sphinx-autobuild==2021.3.14
sphinx-autodoc-typehints==1.18.1
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-spelling==7.3.2
SQLAlchemy==2.0.25
sqlparse==0.5.0
srsly==2.4.8
sshconf==0.2.5
-e git+https://github.com/DLR-RM/stable-baselines3@a6f5049a99a4c21a6f0bcce458ca3306cef310e0#egg=stable_baselines3
starlette==0.27.0
sympy==1.12
syncer==2.0.3
tabulate==0.8.9
tenacity==8.2.3
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.0
tensorflow-addons==0.16.1
tensorflow-estimator==2.4.0
tensorflow-hub==0.12.0
termcolor==1.1.0
terminado==0.9.3
testpath==0.4.4
thinc==8.2.3
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tokenizers==0.15.0
toml==0.10.1
tomli==2.0.1
torch==2.1.2
torchvision==0.16.2
tornado==6.1
tqdm==4.66.1
traitlets==5.0.5
transformers==4.36.2
typeguard==2.13.3
typer==0.9.0
typing-inspect==0.7.1
typing_extensions==4.9.0
ultralytics==8.2.51
ultralytics-thop==2.0.0
uptrace==1.22.0
urllib3==1.26.4
ursina==4.1.1
uvicorn==0.23.2
uvloop==0.19.0
varname==0.10.0
wasabi==1.1.2
watchfiles==0.20.0
wcwidth==0.2.5
weasel==0.3.4
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==2.2.2
widgetsnbextension==3.5.1
wrapt==1.12.1
wsproto==1.2.0
yarl==1.9.4
zipp==3.8.0
